Related to #8 updated the fitness classes  database document to include the necessary functions
